backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "merge-backend"
version = "0.1.0"
description = "FastAPI backend for Merge"
requires-python = ">=3.11"
dependencies = [
  "celery>=5.4,<6.0",
  "cryptography>=42.0,<44.0",
  "fastapi>=0.115,<1.0",
  "httpx>=0.27,<1.0",
  "pg8000>=1.31,<2.0",
  "pydantic-settings>=2.4,<3.0",
  "python-multipart>=0.0.9,<1.0",
  "rapidfuzz>=3.9,<4.0",
  "redis>=5.0,<6.0",
  "sqlalchemy>=2.0,<3.0",
  "tenacity>=8.4,<9.0",
  "uvicorn[standard]>=0.30,<1.0",
  "ytmusicapi>=1.8,<2.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.3,<9.0"]

[project.scripts]
app = "app.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
